주캐 이미지 표시 구현

+++ ERHistoryViewer/app/src/main/java/com/example/erhistoryviewer/CharacterStats.java
@@ -7,6 +7,6 @@ public class CharacterStats {
     int maxKillings;
     int top3;
     int wins;
-    int top3Rate;
-    int averageRank;
+    float top3Rate;
+    float averageRank;
 }
